Implement smooth square movement. Implement GameObject and GamePart. Implement two stage server side movement (for awaited collisions). There is a way to play game with yourself. Implement Integration tests.

+++ src/main/java/ru/mail/park/mechanics/Config.java
@@ -4,7 +4,7 @@ package ru.mail.park.mechanics;
  * Created by Solovyev on 03/11/2016.
  */
 public class Config {
-    public static final float SQUARE_SPEED = 0.2f; //pix per ms
+    public static final double SQUARE_SPEED = 0.2f; //pix per ms
     public static final int PLAYGROUND_WIDTH = 640;
     public static final int PLAYGROUND_HEIGHT = 480;
     public static final int SQUARE_SIZE = 30;
